+++ packages/office-ui-fabric-react/src/components/Dialog/examples/Dialog.Blocking.Example.tsx
@@ -2,18 +2,12 @@ import * as React from 'react';
 import { Dialog, DialogType, DialogFooter } from 'office-ui-fabric-react/lib/Dialog';
 import { PrimaryButton, DefaultButton } from 'office-ui-fabric-react/lib/Button';
 
-export class DialogBlockingExample extends React.Component<
-  {},
-  {
-    hideDialog: boolean;
-  }
-> {
-  constructor(props: {}) {
-    super(props);
-    this.state = {
-      hideDialog: true
-    };
-  }
+export interface IDialogBlockingExampleState {
+  hideDialog: boolean;
+}
+
+export class DialogBlockingExample extends React.Component<{}, IDialogBlockingExampleState> {
+  public state: IDialogBlockingExampleState = { hideDialog: true };
 
   public render() {
     return (
@@ -29,7 +23,7 @@ export class DialogBlockingExample extends React.Component<
           }}
           modalProps={{
             isBlocking: true,
-            containerClassName: 'ms-dialogMainOverride'
+            styles: { main: { maxWidth: 450 } }
           }}
         >
           <DialogFooter>