+++ packages/react-cards/src/components/Card/Card.styles.ts
@@ -1,4 +1,4 @@
-import { getGlobalClassNames } from '@uifabric/styling';
+import { getGlobalClassNames, HighContrastSelector } from '@uifabric/styling';
 import { Depths } from '@uifabric/fluent-theme';
 import { ICardComponent, ICardStylesReturnType, ICardTokenReturnType } from './Card.types';
 
@@ -11,6 +11,7 @@ const baseTokens: ICardComponent['tokens'] = {
   boxShadow: Depths.depth4,
   childrenGap: 12,
   height: 'inherit',
+  highContrastBoxShadow: '0 1.6px 3.6px 0 Highlight, 0 0.3px 0.9px 0 Highlight',
   minHeight: '348px',
   minWidth: '212px',
   maxWidth: '286px',
@@ -24,7 +25,8 @@ const compactTokens: ICardComponent['tokens'] = {
 };
 
 const clickableTokens: ICardComponent['tokens'] = {
-  boxShadowHovered: Depths.depth8
+  boxShadowHovered: Depths.depth8,
+  highContrastBoxShadowHovered: '0 3.2px 7.2px 0 Highlight, 0 0.6px 1.8px 0 Highlight'
 };
 
 export const CardTokens: ICardComponent['tokens'] = (props, theme): ICardTokenReturnType => [
@@ -51,7 +53,15 @@ export const CardStyles: ICardComponent['styles'] = (props, theme, tokens): ICar
 
         selectors: {
           ':hover': {
-            boxShadow: tokens.boxShadowHovered
+            boxShadow: tokens.boxShadowHovered,
+            selectors: {
+              [HighContrastSelector]: {
+                boxShadow: tokens.highContrastBoxShadowHovered
+              }
+            }
+          },
+          [HighContrastSelector]: {
+            boxShadow: tokens.highContrastBoxShadow
           }
         }
       }